pyproject.toml
[project]
name = "financials"
version = "0.1.0"
description = "Financials analysis tool"
readme = "README.md"
authors = [
    { name = "Yacin Bahi", email = "yacin@red64.io" }
]
requires-python = ">=3.11"

dependencies = [
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "streamlit>=1.45.0",
    "unstructured[all-docs]>=0.17.2",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "weaviate-client>=4.4.0",
    "chromadb>=1.0.7",
    "tiktoken>=0.9.0",
]

[project.scripts]
finanalyze = "financials.cli.cli:main"
